Moved the ebpf stuff to its own package under private/underlay.
This way it can be used not only by the router, but also by brload or whatever
needs it. (requires capabilities, though).
