+++ private/underlay/ebpf/portfilter.c
@@ -1,16 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+//
 // Copyright 2025 SCION Association
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
 //go:build ignore
 
@@ -67,4 +57,5 @@ int bpf_port_filter(struct __sk_buff *skb)
   return skb->len;
 }
 
-char __license[] SEC("license") = "Dual MIT/GPL";
+// This program only uses non-gpl_only helpers. So we can use our normal license.
+char __license[] SEC("license") = "Apache-2.0";
